+++ tests/torch2/fx/test_min_max.py
