+++ tests/torch2/fx/test_models.py
@@ -39,12 +39,12 @@ from nncf.experimental.torch.fx.transformations import _get_node_inputs
 from nncf.quantization.advanced_parameters import AdvancedQuantizationParameters
 from tests.cross_fw.shared.paths import TEST_ROOT
 from tests.torch import test_models
-from tests.torch.fx.helpers import get_torch_fx_model
-from tests.torch.fx.test_sanity import count_q_dq
 from tests.torch.test_compressed_graph import check_graph
 from tests.torch.test_models.synthetic import MultiBranchesConnectedModel
 from tests.torch.test_models.synthetic import ShortTransformer
 from tests.torch.test_models.synthetic import YOLO11N_SDPABlock
+from tests.torch2.fx.helpers import get_torch_fx_model
+from tests.torch2.fx.test_sanity import count_q_dq
 
 FX_DIR_NAME = Path("fx")
 FX_QUANTIZED_DIR_NAME = FX_DIR_NAME / "quantized"
