+++ tests/torch2/fx/test_ptq_params.py
@@ -33,10 +33,10 @@ from tests.common.quantization.metatypes import Conv2dTestMetatype
 from tests.common.quantization.metatypes import LinearTestMetatype
 from tests.common.quantization.metatypes import SoftmaxTestMetatype
 from tests.cross_fw.test_templates.test_ptq_params import TemplateTestPTQParams
-from tests.torch.fx.helpers import get_single_conv_nncf_graph
-from tests.torch.fx.helpers import get_torch_fx_model_q_transformed
 from tests.torch.test_models.synthetic import LinearPTQParamsTestModel
 from tests.torch2.function_hook.quantization.helper import get_single_no_weight_matmul_nncf_graph
+from tests.torch2.fx.helpers import get_single_conv_nncf_graph
+from tests.torch2.fx.helpers import get_torch_fx_model_q_transformed
 
 
 def get_hw_patterns(device: TargetDevice = TargetDevice.ANY) -> GraphPattern:
