+++ tests/torch2/fx/test_statistics_caching.py
